Removed gobject dependencies in the main loop, Start also to move things around with a concept of applet and plugins. Will also make the main look more Pygame friendly and remove all independant call to display things, and force it to a fixed framerate (may add an option to change it, it is currently set to 30fps)
Applets are all the apps that run in the UI, like Settings, the music player, the app icon list, etc..
Plugins are all the thing running on top of the UI like the Wifi icon, Battery infos, time, or the brighness.
Config have change from python code to JSON file, with the ability to save the config.

The use of applets or plugin is to remove hard dependencies on things like dbus/wicd/mpd that should be part of applet/plugins and not the main code/loop.